redux refactoring
